# api-docs.yaml
openapi: 3.0.1
info:
  title: iBoard 社區公佈欄(後端Spring boot)
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /users/{userId}:
    get:
      tags:
      - 住戶 CRUD
      summary: 取得住戶資料
      description: 透過userId取得住戶資料並回傳前端
      operationId: getUserById
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    put:
      tags:
      - 住戶 CRUD
      summary: 修改住戶資料
      description: 透過 userId 修改住戶資料
      operationId: updateUser
      parameters:
      - name: userId
        in: path
        description: 住戶user_id
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        description: 請求參數列表
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        required: true
      responses:
        "404":
          description: 無此 住戶 ID，無法執行修改功能
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    delete:
      tags:
      - 住戶 CRUD
      summary: 刪除住戶資料
      description: 透過 userId 刪除住戶資料
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        description: 住戶user_id
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: 成功
...
  /users:
    get:
      tags:
      - 住戶 CRUD
      summary: 住戶資料列表(模糊查詢+分頁+排序)
      description: 可模糊查詢姓名，依照前端請求orderBy指定項目排序(預設為創建資料時間)，依照前端請求sort指定排序順序(DESC由大到小排序)，依照前端請求limit限制每頁筆數(預設5)，依照前端請求offset設定跳過筆數(預設0，代表第一頁)。回傳總數count讓前端計算分頁。
      operationId: getUsers
      parameters:
      - name: search
        in: query
        description: 關鍵字搜尋(住戶姓名)
        required: false
        schema:
          type: string
      - name: orderBy
        in: query
        description: 根據所選欄位排序
        required: false
        schema:
          type: string
          default: created_date
      - name: sort
        in: query
        description: 排序方式
        required: false
        schema:
          type: string
          default: desc
      - name: limit
        in: query
        description: 限制資料筆數
        required: false
        schema:
          maximum: 1000
          minimum: 0
          type: integer
          format: int32
          default: 5
      - name: offset
        in: query
        description: 跳過資料筆數
        required: false
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags:
      - 住戶 CRUD
      summary: 新增住戶資料
      description: 新增住戶資料，完成後回傳新增的住戶資訊
      operationId: createUser
      requestBody:
        description: 請求參數列表
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
...
components:
  schemas:
    UserRequest:
      required:
      - userAddress
      - userName
      - userTel
      type: object
      properties:
        userFaceId:
          type: string
          description: 住戶 ID
        userName:
          type: string
          description: 住戶姓名
        userTel:
          type: string
          description: 住戶聯絡電話
        userAddress:
          type: string
          description: 住戶住址
    User:
      type: object
      properties:
        userId:
          type: integer
          description: 住戶 ID
          format: int32
        userFaceId:
          type: string
          description: 住戶 人臉辨識 ID
        userName:
          type: string
          description: 住戶姓名
        userTel:
          type: string
          description: 住戶聯絡電話
        userAddress:
          type: string
          description: 住戶地址
        createdDate:
          type: string
          description: 住戶資料被建立時間
          format: date-time
        lastModifiedDate:
          type: string
          description: 住戶資料最後修改時間
          format: date-time
...